Add files via upload
Where you need to start modifying the main program to work with the wounded soldier giving out quests is at line 350. In the file CSC1061_Final_Rivera_Ardrey.cpp . It ends at line 401. Thank you so much for finishing the program, and I am regretful I cannot do it myself, or with you. :(